Add ShipmentLine nested model
This adds coercion and nested models. The motivation here being that
outside of this gem we should accept a single properly structured hash,
and that hash can contain many models and will need to in order to
provide all the correct data in order to do something like create a
shipment inside of Snap. To try and easy the use of developing and using
this gem, on our side we will automatically look for specific nested
models and initialize them.