Add errors!
This adds a block to snoop into responses and see if they have
encountered a common error. This is helpful for consumers to control the
flow of the code in certain situations. This error is especially is very
odd to identify. This is also not something we could have prevented by
snooping the request. Snap is stateful and we don't know the state
pre-flight.